requirements.txt
streamlit>=0.88
streamlit-drawable-canvas>=0.8
svgpathtools
svgwrite
numpy
pillow
opencv-python
pyimgur
dropbox